Major Update
Added UICollectionView to Home screen with customized UICollectionViewCell's and attributes.

Added some polish to the hamburger menu, and removed the old tableview and replaced with 2 buttons with custom attributes

Added some basic functionality to the UICollectionViewCell's so they act according to their meaning. Add new house will add a new house, and clicking on a existing house is currently not hooked up.